Updated control.sh to allow for symlinking to the script.
The patch allows a symlink to the control.sh to resolve to the actual location of the cronicle control script. Useful when you are adding a symlink to control.sh into another folder on your system (e.g. /etc/init.d) and want it to still find the process .pid file in the actual location, instead of looking in the folder where the link is.